adapted structure so that templates take lists of (layers of) parameters.
