+++ tests/test_templates_embeddings.py
